Move synthetic PSD creating out to it's own file in the module
